refactor: UseCase constructor 수정
