+++ PayKids/domain/src/main/java/com/paykids/domain/usecase/auth/KakaoAuthUseCase.kt
@@ -1,12 +1,10 @@
-package com.paykids.domain.usecase
+package com.paykids.domain.usecase.auth
 
 import com.paykids.domain.model.SignInInfo
 import com.paykids.domain.repository.KakaoAuthRepository
 import javax.inject.Inject
 
-class KakaoAuthUseCase @Inject constructor(
-    private val repository: KakaoAuthRepository
-) {
+class KakaoAuthUseCase @Inject constructor(private val repository: KakaoAuthRepository) {
     suspend operator fun invoke(): SignInInfo {
         return repository.signInWithKakao()
     }
